msm: camera: actuator: remove unneeded copy_from_user operation for init_settings
the allocation of enough space for init_settings is sometimes failing
due to a mysterious bug in memory managment.
init_settings is only being read in the succeeding functions, and
copying a read-only structure from user to kernel holds little sense.
Avoid the allocation bug by simply not copying the structure over to
kernel and accessing the information from the user structure directly.